+++ src/compiler/parser/mod.rs
@@ -18,8 +18,10 @@ mod lines_5000_5500;
 pub use lines_5000_5500::*;
 mod lines_5500_6000;
 pub use lines_5500_6000::*;
-mod lines_6000_7000;
-pub use lines_6000_7000::*;
+mod lines_6000_6500;
+pub use lines_6000_6500::*;
+mod lines_6500_7000;
+pub use lines_6500_7000::*;
 mod lines_7000_8000;
 pub use lines_7000_8000::*;
 mod lines_8000_9000;
